Combine semver bumps and publish into single jobs
- Add rubygems API key to publish job
- Generate `version` file on the fly rather than checking in


[#136975857](https://www.pivotaltracker.com/story/show/136975857)

+++ ci/tasks/publish-gem.sh
@@ -2,25 +2,22 @@
 
 set -e
 
-check_param() {
-  local name=$1
-  local value=$(eval echo '$'$name)
-  if [ "$value" == 'replace-me' ]; then
-    echo "environment variable $name must be set"
-    exit 1
-  fi
-}
+: ${RUBYGEMS_API_KEY:?}
+: ${RUBY_VERSION:="2.3.1"}
 
 version="$( cat version-semver/version )"
-
-check_param RUBY_VERSION
 
 source /etc/profile.d/chruby.sh
 chruby $RUBY_VERSION
 
+cat << EOF > ~/.gem/credentials
+---
+:rubygems_api_key: ${RUBYGEMS_API_KEY}
+EOF
+
 pushd bosh-src
-    git clean -xfd
-    echo "${version}" > version
-    gem build *.gemspec
-    gem push bosh_cpi-*.gem
+  git clean -xfd
+  echo "${version}" > version
+  gem build *.gemspec
+  gem push bosh_cpi-*.gem
 popd
